Use HTTP Post in backend requests with "corpus" parameter
Use the HTTP POST method for Korp backend requests with the "corpus"
parameter, to avoid "URI too long" errors when a large number of corpora
are selected.

This fixes #214, except for the JSON buttons that are still links
containing the backend parameters in the URL, so they will not work with
a large number of corpora.