Fix #222: Do not duplicate parameters in JSON button URLs
model.js: BaseProxy.makeUrlWithParams: If the base URL already contains
URL parameters, do not add them from data. When the function is called
for a GET URL, the base URL already contains parameters, which
previously resulted in duplicating the parameters, which in turn
resulted in ValueErrors from the backend when trying to get the JSON.
This fixes #222.