Fix #201: KWIC: Select first match after every search
KWIC result: Select the first matched word in the first hit and show its
information in the sidebar after every search and not only after the
initial search. Previously, further searches did not select the word and
the sidebar showed information for the word selected in the previous
search result. This fixes #201.