fix: improve handling of missing stats_reduce
There is a guard in StatsProxy but it used CorpusListing.getReduceAttrs() which didn't take `display_type: hidden` into account. Now it does, and this commit also changes some UX around the error message.

There's an overflow problem when opening the groupby dropdown, but I
didn't manage to fix it right now.

Fixes #314